requirements.txt
langgraph
langchain
langchain-groq
pydantic
matplotlib
gradio 
pypdf 
beautifulsoup4
langchain-community
python-dotenv
markdown
xhtml2pdf